+++ calcbutton.h
@@ -6,34 +6,34 @@
 class CalcButton : public QPushButton {
     public:
         std::string operation = "?";
-        CalcButton(Calculator<int(int, int)>&, const QString &, QWidget *);
+        CalcButton(Calculator<int64_t(int64_t, int64_t)>&, const QString &, QWidget *);
         virtual void on_click() = 0;
     protected:
-        Calculator<int(int,int)> &calculator;
+        Calculator<int64_t(int64_t,int64_t)> &calculator;
 };
 
 class DigitButton: public CalcButton {
     public:
         uint8_t digit;
-        DigitButton(Calculator<int(int, int)>&, QWidget*, uint8_t digit);
+        DigitButton(Calculator<int64_t(int64_t, int64_t)>&, QWidget*, uint8_t digit);
         void on_click();
 };
 
 class PlusButton: public CalcButton {
     public:
-        PlusButton(Calculator<int(int, int)>&, QWidget*);
+        PlusButton(Calculator<int64_t(int64_t, int64_t)>&, QWidget*);
         void on_click();
 };
 
 class EqualButton: public CalcButton {
     public:
-        EqualButton(Calculator<int(int, int)>&, QWidget*);
+        EqualButton(Calculator<int64_t(int64_t, int64_t)>&, QWidget*);
         void on_click();
 };
 
 class ClearButton: public CalcButton {
     public:
-        ClearButton(Calculator<int(int, int)>&, QWidget*);
+        ClearButton(Calculator<int64_t(int64_t, int64_t)>&, QWidget*);
         void on_click();
 };
